middleware/requirements.txt
crate
pandas
numpy
scikit-learn
psycopg2-binary
skforecast